apply datafusion change

+++ src/service/search/tantivy/puffin/reader.rs
@@ -73,7 +73,7 @@ impl PuffinBytesReader {
             return Ok(());
         }
 
-        if (self.source.size as u64) < MIN_FILE_SIZE {
+        if self.source.size < MIN_FILE_SIZE {
             return Err(anyhow!(
                 "Unexpected bytes size: minimal size {} vs actual size {}",
                 MIN_FILE_SIZE,
@@ -218,7 +218,7 @@ impl PuffinFooterBytesReader {
             .map_or(MAGIC_SIZE, |blob| blob.offset + blob.length);
         let footer_size = MAGIC_SIZE + self.payload_size + FOOTER_SIZE;
         ensure!(
-            payload_ends_at == (self.source.size as u64 - footer_size),
+            payload_ends_at == (self.source.size - footer_size),
             anyhow!("Payload chunk offset mismatch")
         );
         Ok(())
